Added more randomisation to the ascii sum experiment to make it harder to predict. Also added a float cast in the Neural Network code.
- The value associated to a character now goes through a hash function to pseudo-randomise it.
- The neural network still seems to be doing very well on this problem.

+++ src/neural_network/neural_network.cpp
@@ -91,7 +91,7 @@ float NeuralNetwork::get_total_loss(std::vector<std::vector<float>> inputs, std:
 	for (size_t i = 0; i < inputs.size(); i++) {
 		loss += mean_squared_error(this->get_prediction(inputs[i]), targets[i]);
 	}
-	return loss / inputs.size();
+	return loss / static_cast<float>(inputs.size());
 }
 
 void NeuralNetwork::train(std::vector<std::vector<float>>& inputs, std::vector<std::vector<float>>& targets, int nb_epochs,
